+++ examples/resources/scalr_hook_environment_link/import.sh
@@ -1,0 +1,1 @@
+scalr_hook_environment_link.example hkenv-xxxxxxxxxx