+++ internal/provider/provider.go
@@ -154,6 +154,7 @@ func (p *scalrProvider) Resources(_ context.Context) []func() resource.Resource 
 		newWorkspaceResource,
 		newIntegrationInfracostResource,
 		newHookResource,
+		newHookEnvironmentLinkResource,
 	}
 }