+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - **New resource:** `scalr_hook` ([#403](https://github.com/Scalr/terraform-provider-scalr/pull/403)
+- **New resource:** `scalr_hook_environment_link` ([#403](https://github.com/Scalr/terraform-provider-scalr/pull/403)
 - **New data source:** `scalr_hook` ([#403](https://github.com/Scalr/terraform-provider-scalr/pull/403)
 
 ## [2.6.0] - 2025-03-07